% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backups_info-class.R
\docType{class}
\name{backups_info-class}
\alias{backups_info-class}
\title{Extract Backups Metadata}
\description{
SQL files produced by \code{\link[=release_project]{release_project()}} contain metadata that can be
extracted into a data.frame.
}
