% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backups_info.R
\name{backups_info}
\alias{backups_info}
\alias{read_info}
\title{Display Backup Metadata}
\usage{
read_info(sql)

backups_info(backup_path = ".")
}
\arguments{
\item{sql}{A character value with the path to a SQL file containing
metadata.}

\item{backup_path}{A character value with the path to a folder containing
SQL files.}
}
\value{
A data frame of class \linkS4class{backups_info}.
}
\description{
Extracting metadata from backups produced by \code{\link[=release_project]{release_project()}} can be
displayed in a data frame of class \linkS4class{backups_info}.

Metadata from single SQL file will be extracted by \code{read_info()}, while
\code{backups_info()} will extract metadata from multiple SQL files.
}
