% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_db.R
\name{build_db}
\alias{build_db}
\title{Restore a Database from a released backup}
\usage{
build_db(dbname, user, backup_path, release = "last", ...)
}
\arguments{
\item{dbname}{A character value with the name of the database where to
restore. Note that this database should be created in advance, otherwise
you will retrieve an error message.}

\item{user}{A character value with the name of the user.}

\item{backup_path}{A character value with the path to the folder containing
backups.}

\item{release}{An integer value indicating the sort order according to
\code{\link[=backups_info]{backups_info()}}. The keywords "first" and "last" are also accepted.
The default value is "last".}

\item{...}{Further arguments passed to \code{\link[divDB:do_backup]{divDB::do_restore()}}.}
}
\description{
Releases are a collection of backups generated by \code{\link[=release_project]{release_project()}}.
After displaying them with \code{\link[=backups_info]{backups_info()}}, you may desire to restore a
determined version of the database.
}
