% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_project.R
\name{init_project}
\alias{init_project}
\title{Initialize a new project for database updates}
\usage{
init_project(
  dbname,
  user,
  path,
  main_script = "main-script",
  remarks = "",
  rs_project = FALSE,
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{dbname}{A character value with the name of the database to be updated.}

\item{user}{A character value with the name of the user connecting the}

\item{path}{A character value indicating the path for the new project.}

\item{main_script}{A character value with the name of the template used as
main script in the project.}

\item{remarks}{A character value describing the project.}

\item{rs_project}{A logical value indicating whether an R Studio project
should be initialized or not. This may be useful to proceed with the
ETL workflow in an own R Studio project.}

\item{overwrite}{A logical value indicating whether an existing homonymous
project should be overwritten or not.}

\item{...}{Further arguments passed to \code{\link[divDB:do_backup]{divDB::do_backup()}} and
\code{\link[divDB:connect_db]{divDB::connect_db()}} (for instance 'host' and 'port').}
}
\description{
A project folder will contain all necessary data and scripts needed to update
a database into a new version. This strategy will allow to rebuild the
database in a preferential version and document at the same time all the
process done on the way.
}
