% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-methods.R
\docType{methods}
\name{print-methods}
\alias{print-methods}
\alias{print.backups_info}
\alias{print,backups_info-method}
\title{Print Methods}
\usage{
\method{print}{backups_info}(x, cols, ...)
}
\arguments{
\item{x}{An object of class \linkS4class{backups_info}.}

\item{cols}{A character vector with the names of columns selected for the
printed output.}

\item{...}{Further arguments passed to next method.}
}
\description{
Print methods for objects of class \linkS4class{backups_info}.
}
