% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_project.R
\name{init_project}
\alias{init_project}
\title{Initialize a new project for database updates}
\usage{
init_project(
  dbname,
  user,
  path,
  remarks = "",
  restore_from,
  r_script = TRUE,
  rs_project = FALSE,
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{dbname}{A character value with the name of the database to be updated.}

\item{user}{A character value with the name of the user connecting the}

\item{path}{A character value indicating the path for the new project.}

\item{remarks}{A character value describing the project.}

\item{restore_from}{A character value. The path to the acummulated backups.
If not provided a warning as reminder will be retrieved.}

\item{r_script}{A logical value indicating whether a template R script
should be included in the project or not. The default is 'TRUE'.}

\item{rs_project}{A logical value indicating whether an R Studio project
should be initialized or not. This may be useful to proceed with the
ETL workflow in an own R Studio project. The default is 'FALSE'.}

\item{overwrite}{A logical value indicating whether an existing homonymous
project should be overwritten or not.}

\item{...}{Further arguments passed to \code{\link[divDB:connect_db]{divDB::connect_db()}},
\code{\link[divDB:do_backup]{divDB::do_backup()}} and \code{\link[divDB:do_backup]{divDB::do_restore()}} (for instance 'host' and
'port').}
}
\description{
A project folder will contain all necessary data and scripts needed to update
a database into a new version. This strategy will allow to rebuild the
database in a preferential version and document at the same time all the
process done on the way.
}
