% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release_project.R
\name{release_project}
\alias{release_project}
\title{Pack an update project for release}
\usage{
release_project(
  project_path,
  backup_path,
  vault,
  release_sql,
  f_timestamp = "\%Y\%m\%d-\%H\%M",
  restore = TRUE,
  delete_project,
  version_table,
  ...
)
}
\arguments{
\item{project_path}{A character value indicating the path to the project
folder. Note that the project folder must be created by \code{\link[=init_project]{init_project()}}.}

\item{backup_path}{A character value indicating the path to the collection
of backup files (usually a sql and a backup file with timestamp).}

\item{vault}{A character value providing the path to the folder, where copies
of projects as zip files will be stored.}

\item{release_sql}{The name of the sql file in the project folder used to
update the database.}

\item{f_timestamp}{A character value indicating the format used for
timestamp. It has to be the same as the one used by \code{\link[=init_project]{init_project()}}.}

\item{restore}{A logical value indicating whether the database should be
first restored from last backup, which is strictly recommended and the
default setting.}

\item{delete_project}{A logical value indicating whether the used folder
project should be deleted after release or not. If not provided, it will
be interactivelly prompted.}

\item{version_table}{A character value with the name of the schema and the
table containing the versions record. Not necessary if this table
is missing in the database.}

\item{...}{Further arguments passed to \code{\link[divDB:do_backup]{divDB::do_backup()}} and
\code{\link[divDB:do_backup]{divDB::do_restore()}} (for instance 'host' and 'port').}
}
\description{
A release project includes an initial backup, a main script and all necessary
data to update a database starting by the backup.

This function will produce a zip file including all the project's content
using the name of the database and a time stamp.
}
